Proxies to MFE servers based on `/mfe{1,2}` paths.
I initially tried making `ProxyService` scoped to the request and injecting the request/response pair there, however Nest can't seem to resolve the dependencies in that context for reasons which escape me. `@Req()` fails, but `@Inject(REQUEST)` succeeds. Except there is no `@Inject(RESPONSE)` corollary, so we need `@Res()`. Not sure what's going on here, but the easiest solution is to inject in the controller and pass the request/response through.