Hard navigate on router match failure.
This is a bit of a hack as it requires the router to fail to navigate, catch the error, let it fail and log to the console, and then perform a hard navigation. This isn't ideal, as there is no way to synchronously catch the error and suppress it without redirecting to another soft navigation, which we can't do here.