Switches hard navigation strategy to a 404 route guard.
The `NO_MATCH` error handler _almost_ worked, however routing from MFE2 to `/` would fail and render an empty route. Not sure exactly what's going on there, but it seems like Angular Router expects a '' route to exist and its absence does not trigger `NO_MATCH`.

This approach uses a 404 route which is only triggered when all other routes fail to match. It abuses as `canActivate` guard to force a hard navigation whenever it is triggered. This seems to handle a hard navigation to `/` correctly.