Adds a not found page for routes outside all MFEs.
This is tricky because each MFE only knows its own routes. Therefore if the user navigates from MFE1 to `/foo`, MFE1 only knows that it doesn't own `/foo`. It does *not* know if any MFEs claim `/foo`. Therefore it can't know whether this navigation will 404.

The proxy server is the one with knowledge of all the MFE routes, so it knows whether a navigation is claimed by an MFE (MFEs themselves might have 404 routes, which the proxy server doesn't know). When this happens, the proxy server redirects to MFE1 which effectively "claims" all routes not claimed by any other MFE.

This then presents a problem, how does MFE1 know whether to render or redirect a 404 route? Essentially it renders any initial hard-navigation, since this can only happen from the proxy server, which would only redirect traffic to MFE1 if no other MFE owned the current route. If the user performs a client-side navigation to any route not known to MFE1, then it performs a hard navigation and lets the proxy server route to the correct MFE (which is either not MFE1, or a 404 route).

This creates some interesting behavior:
1. The router is now stateful and does different things between the initial navigation and all other navigations.
2. Navigating to a not found route which is not owned by any MFE is always a hard navigation, even if this navigation originates from MFE1 and will ultimately be redirected back to MFE1.
    * I think this is ok because realistically it is very rare for a 404 route to link to _another_ 404 route.
3. Navigations between known routes and not found routes within a single MFE still works as normal. It's only 404 routes outside of any MFE which encounter this unique edge case behavior.

This could be addressed by giving MFE1 knowledge of all other routes in the system, this would allow it to disambiguate navigations to other MFEs from 404 routes, but I didn't want to introduce that knowledge. Instead, only the proxy server has knowledge of full app routing. Each individual MFE only controls their own 404 routes, with one aribtrary MFE (MFE1 in this case) receiving all 404 routes not claimed by any MFEs.