Moves `NavigationTracker` to common library and adds it to MFE2.
I initially thought this was only needed for MFE1 to handle navigations to unclaimed routes. However it turns out this is useful for MFE2 running on a devserver. Specifically it needs to handle 404 navigations and also will frequently be loaded at `/` by the developer when first loading the app. This naturally doesn't work, but would cause an infinite navigation loop to `/`. With `NavigationTracker` we break out of the loop and throw an error pointing the user at the correct "home page" for MFE2.