Removes `servePath` and sets `baseHref: "/"` for devserving.
The README has more details, but basically we need `baseHref: "/"` for devservers in order for MFE1 to correctly render the `/` route, otherwise it will 302 redirect to `/mfe1`. This means we have to drop `servePath` as well to load the assets correctly. This diverges the location of assets such as `main.js` and `styles.css` between the devserver and production server, but either way Angular loads it correctly so it should be fine?